% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Adam}
\alias{Adam}
\alias{adam_optimization}
\title{Adam Optimization}
\usage{
adam_optimization(X, y, lr, beta1, beta2, epsilon, num_iters)
}
\arguments{
\item{X}{The design matrix of size M*N}

\item{y}{The dependent vector of size M*1}

\item{lr}{The learning rate}

\item{beta1}{Exponential decay rate for the first moment estimate}

\item{beta2}{Exponential decay rate for the second moment estimate}

\item{epsilon}{Small constant to avoid division by zero}

\item{num_iters}{The number of iterations}
}
\value{
The vector of optimal parameter values
}
\description{
Adam optimization algorithm
}
\examples{
\dontrun{
X <- matrix(runif(1000), ncol = 20)
y <- 2*rbinom(50,1,0.5)-1
lr <- 0.001
beta1 <- 0.9
beta2 <- 0.999
epsilon <- 1e-8
num_iters <- 100
theta_adam <- adam_optimization(X, y, lr, beta1, beta2, epsilon, num_iters)
}
}
