\name{BA23229018-package}
\alias{BA23229018-package}
\alias{BA23229018}
\docType{package}
\title{Final project for 'Statistical Computing' course 2023Aut}
\description{Several R and Rcpp function concerning Deep Learning are included in this package for illustration. }
\author{
Yuanhao Pu
Maintainer: Yuanhao Pu <puyuanhao@mail.ustc.edu.com>}
\keyword{package}
