% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GD}
\alias{GD}
\alias{gradient_descent}
\title{Gradient Descent}
\usage{
gradient_descent(X, y, lr, num_iters)
}
\arguments{
\item{X}{The design matrix of size M*N}

\item{y}{The dependent vector of size M*1}

\item{lr}{The learning rate}

\item{num_iters}{The number of iterations}
}
\value{
The vector of optimal parameter values
}
\description{
A simple implementation of Gradient Descent optimization, here we use a linear regression model for example. Any other type of loss function is reachable.
}
\examples{
\dontrun{
X <- matrix(runif(1000), ncol = 20)
y <- 2*rbinom(50,1,0.5)-1
lr <- 0.001
num_iters <- 100
theta <- gradient_descent(X, y, lr, num_iters)
}
}
