% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{MinibatchGD}
\alias{MinibatchGD}
\alias{minibatch_gradient_descent}
\title{Mini-Batch Gradient Descent}
\usage{
minibatch_gradient_descent(X, y, lr, num_iters, batch_size)
}
\arguments{
\item{X}{The design matrix of size M*N}

\item{y}{The dependent vector of size M*1}

\item{lr}{The learning rate}

\item{num_iters}{The number of iterations}

\item{batch_size}{The size of mini-batch}
}
\value{
The vector of optimal parameter values
}
\description{
Mini-Batch Gradient Descent optimization algorithm
}
\examples{
\dontrun{
X <- matrix(runif(1000), ncol = 20)
y <- 2*rbinom(50,1,0.5)-1
lr <- 0.0001
num_iters <- 100
batch_size <- 10
theta_minibatch <- minibatch_gradient_descent(X, y, lr, num_iters, batch_size)
}
}
