% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NeuralNetwork.R
\name{mlp_train}
\alias{mlp_train}
\title{A training process of NeuralNetwork}
\usage{
mlp_train(X, y, hidden_size, lr, epochs, lambda = 0.01)
}
\arguments{
\item{X}{the feature dataset}

\item{y}{the target label}

\item{hidden_size}{the width of hidden layer}

\item{lr}{learning rate}

\item{epochs}{number of epochs to be trained}

\item{lambda}{regularization coefficient}
}
\value{
The loss graph
}
\description{
A training process of NeuralNetwork, including BP algorithm of a simple MLP model
}
\examples{
\dontrun{
train_inputs <- matrix(runif(1000), ncol = 20, byrow = TRUE)
train_targets <- 2*rbinom(50, 1, 0.5)-1
mlp_train(train_inputs, train_targets, hidden_size = 8, lr = 0.001, epochs = 50, lambda = 0.001)
}
}
