% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SGD}
\alias{SGD}
\alias{stochastic_gradient_descent}
\title{Stochastic Gradient Descent}
\usage{
stochastic_gradient_descent(X, y, lr, num_iters)
}
\arguments{
\item{X}{The design matrix of size M*N}

\item{y}{The dependent vector of size M*1}

\item{lr}{The learning rate}

\item{num_iters}{The number of iterations}
}
\value{
The vector of optimal parameter values
}
\description{
Stochastic Gradient Descent optimization algorithm
}
\examples{
\dontrun{
X <- matrix(runif(1000), ncol = 20)
y <- 2*rbinom(50,1,0.5)-1
lr <- 0.001
num_iters <- 100
theta_sgd <- stochastic_gradient_descent(X, y, lr, num_iters)
}
}
