% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SGDM}
\alias{SGDM}
\alias{sgd_momentum}
\title{Stochastic Gradient Descent with Momentum}
\usage{
sgd_momentum(X, y, lr, momentum, num_iters)
}
\arguments{
\item{X}{The design matrix of size M*N}

\item{y}{The dependent vector of size M*1}

\item{lr}{The learning rate}

\item{momentum}{The momentum parameter}

\item{num_iters}{The number of iterations}
}
\value{
The vector of optimal parameter values
}
\description{
Stochastic Gradient Descent optimization algorithm with momentum
}
\examples{
\dontrun{
X <- matrix(runif(1000), ncol = 20)
y <- 2*rbinom(50,1,0.5)-1
lr <- 0.0001
momentum <- 0.9
num_iters <- 100
theta_sgd_momentum <- sgd_momentum(X, y, lr, momentum, num_iters)
}
}
